0_INFRASTRUCTURE_CORE/requirements.txt
psutil
pyyaml
websocket-client
pandas
numpy
kiteconnect
requests
flask
flask-cors
flask-socketio
python-socketio
eventlet
